refactor(extensions): move _replace_image_tag to compose_transformer module
`_replace_image_tag` was added to `commands/publish.py` in the
preceding namespace-preservation commit, but its callers will expand
to include `ComposeTransformer.transform_service` (the fallback path
when an extension has no `docker-compose.appgarden.yml`) and
`ImageBuilder.build`. To keep the import graph tidy (lib modules
shouldn't import from `commands/`), move the helper to
`compose_transformer.py` and update callers.

Pure refactor: no behavior change. `publish.py` now imports the
helper from `compose_transformer`. `TestReplaceImageTag` import path
updated.
